Prepare removal of `Specifications`
`Specifications` is deprecated since 2.0. This patch prepares the
code for removing it completely.

 * Moves static helper code into a private helper class
 * Removes references to `Specifications` from `Specification`
